refactor(infra): move adapters to infra/adapters for better separation of infrastructure concerns
Moved all files and folders under `@/adapters` to `@/infra/adapters` to better reflect their role as infrastructure layer implementations.

Moved:
- Cart adapters:
  - `@/adapters/cart/...` ➜ `@/infra/adapters/cart/...`
- Product adapters:
  - `@/adapters/product/...` ➜ `@/infra/adapters/product/...`

This structural refactor improves the modularity and domain-driven design of the project by keeping application-specific logic separate from infrastructure-related code. All related imports have been updated accordingly.
