--- adapters/product/product-repository.adapter.ts
@@ -1,3 +1,0 @@
-import { prismaProductRepositoryAdapter } from "@/adapters/product/prisma/prisma-product-repository.adapter";
-
-export const productRepositoryAdapter = prismaProductRepositoryAdapter;
