+++ infra/adapters/cart/cart-repository.adapter.ts
@@ -1,6 +1,6 @@
 import { CartRepository } from "@/domain/repositories/cart.repository";
 
-import { prismaCartRepositoryAdapter } from "@/adapters/cart/prisma/prisma-cart-repository.adapter";
+import { prismaCartRepositoryAdapter } from "@/infra/adapters/cart/prisma/prisma-cart-repository.adapter";
 
 export const cartRepositoryAdapter: CartRepository =
   prismaCartRepositoryAdapter;
