+++ infra/adapters/cart/prisma/get-prisma-cart-by-session-cart-id-service.adapter.ts
@@ -1,6 +1,6 @@
 import { CartEntity } from "@/domain/entities/cart.entity";
 import { getPrismaCartBySessionCartIdService } from "@/infra/services/cart/get-prisma-cart-by-session-cart-id.service";
-import { handleGetPrismaCart } from "@/adapters/cart/prisma/handle-get-prisma-cart";
+import { handleGetPrismaCart } from "@/infra/adapters/cart/prisma/handle-get-prisma-cart";
 
 export async function getPrismaCartBySessionCartIdServiceAdapter(
   sessionCartId: string
