+++ infra/adapters/product/memory/get-memory-product-by-slug-service.adapter.ts
@@ -1,6 +1,6 @@
 import { ProductEntity } from "@/domain/entities/product.entity";
-import { productsInMemory } from "@/adapters/product/memory/db";
-import { mapMemoryProductToDomainProduct } from "@/adapters/product/memory/map-memory-product-to-domain-product";
+import { productsInMemory } from "@/infra/adapters/product/memory/db";
+import { mapMemoryProductToDomainProduct } from "@/infra/adapters/product/memory/map-memory-product-to-domain-product";
 
 export async function getMemoryProductBySlugServiceAdapter(
   slug: string
