+++ infra/adapters/product/prisma/map-prisma-product-to-domain-product.ts
