+++ infra/adapters/product/product-repository.adapter.ts
@@ -1,0 +1,3 @@
+import { prismaProductRepositoryAdapter } from "@/infra/adapters/product/prisma/prisma-product-repository.adapter";
+
+export const productRepositoryAdapter = prismaProductRepositoryAdapter;
